refactor(core): modularize cache provider architecture
