+++ src/lan/server.ts
@@ -1,18 +1,41 @@
 import * as crypto from "crypto";
 import * as fs from "fs";
-import type { IncomingMessage, ServerResponse } from "http";
 import * as https from "https";
 import * as path from "path";
 import { pipeline } from "stream/promises";
 
-import {
-  acquireEntryLock,
-  releaseEntryLock,
-  type EntryLock,
-} from "../cache/lock";
+import { acquireEntryLock, releaseEntryLock } from "../cache/lock";
 import type { CachePlatform } from "../cache/paths";
-import { NonceReplayGuard, verifyAuthHeaders } from "./auth";
-import type { LanAuthorizedClient } from "./types";
+import {
+  authFields,
+  awaitBeforeDeadline,
+  ensureIncomingDirectory,
+  ensureTransferLocksDirectory,
+  framingIsSafe,
+  isStrictRequestTarget,
+  LanOperationTimeoutError,
+  parseContentLength,
+  readBody,
+  rejectRequest,
+  sendEmpty,
+  sendJson,
+  singleHeader,
+  spoolBody,
+} from "./server-http";
+import type {
+  LanServerHandle,
+  LanServerOptions,
+  PreparedLanEntry,
+} from "./server-types";
+export {
+  createLanServerAuthenticator,
+  type LanAuthenticationRequest,
+  type LanAuthenticatedClient,
+  type LanCapability,
+  type LanServerHandle,
+  type LanServerOptions,
+  type PreparedLanEntry,
+} from "./server-types";
 
 const EMPTY_SHA256 = crypto.createHash("sha256").digest("hex");
 const MAX_PAIR_BODY_BYTES = 16 * 1024;
@@ -20,371 +43,6 @@ const MAX_ENTRY_BYTES = 100 * 1024 ** 3 + 128 * 1024;
 const MAX_OPERATION_TIMEOUT_MS = 45_000;
 const ENTRY_ROUTE = /^\/v1\/entries\/(ios|android)\/([a-f0-9]{64})$/;
 const SHA256 = /^[a-f0-9]{64}$/;
-
-export type LanCapability = "read" | "write";
-
-export type LanAuthenticatedClient = {
-  clientId: string;
-  capabilities: readonly LanCapability[];
-};
-
-export type LanAuthenticationRequest = {
-  clientId: string;
-  timestamp: string;
-  nonce: string;
-  contentSha256: string;
-  signature: string;
-  contentLength: number;
-  method: string;
-  pathname: string;
-  allowPendingPairing: boolean;
-};
-
-export type PreparedLanEntry = {
-  packagePath: string;
-  sizeBytes: number;
-  sha256: string;
-  cleanup?: () => void | Promise<void>;
-};
-
-export type LanServerOptions = {
-  host?: string;
-  port?: number;
-  certificatePem: string;
-  privateKeyPem: string;
-  serverId: string;
-  allowWrite?: boolean;
-  incomingDirectory: string;
-  transferLocksRoot?: string;
-  authenticate: (
-    request: LanAuthenticationRequest
-  ) => LanAuthenticatedClient | null | Promise<LanAuthenticatedClient | null>;
-  pair?: (
-    body: unknown,
-    context: { remoteAddress: string }
-  ) => unknown | Promise<unknown>;
-  acknowledgePairing?: (
-    pairingId: string,
-    client: LanAuthenticatedClient
-  ) => unknown | Promise<unknown>;
-  prepareEntry: (
-    platform: CachePlatform,
-    entryId: string,
-    deadlineMs: number
-  ) => PreparedLanEntry | null | Promise<PreparedLanEntry | null>;
-  importEntry: (
-    platform: CachePlatform,
-    entryId: string,
-    packagePath: string,
-    transferLock: EntryLock | undefined,
-    deadlineMs: number
-  ) =>
-    | "created"
-    | "existing"
-    | "conflict"
-    | Promise<"created" | "existing" | "conflict">;
-  globalTransferLimit?: number;
-  perClientTransferLimit?: number;
-  requestInactivityMs?: number;
-  operationTimeoutMs?: number;
-};
-
-export const createLanServerAuthenticator = (options: {
-  findClient: (
-    clientId: string
-  ) => LanAuthorizedClient | null | Promise<LanAuthorizedClient | null>;
-  allowPendingClient?: (
-    client: LanAuthorizedClient,
-    request: LanAuthenticationRequest
-  ) => boolean | Promise<boolean>;
-  replayGuard?: NonceReplayGuard;
-}): LanServerOptions["authenticate"] => {
-  const replayGuard = options.replayGuard ?? new NonceReplayGuard();
-  return async (request) => {
-    const client = await options.findClient(request.clientId);
-    if (!client) return null;
-    const statusAllowed =
-      client.status === "active" ||
-      (request.allowPendingPairing &&
-        client.status === "pending" &&
-        options.allowPendingClient &&
-        (await options.allowPendingClient(client, request)));
-    if (!statusAllowed) return null;
-    try {
-      verifyAuthHeaders({
-        headers: {
-          clientId: request.clientId,
-          timestamp: request.timestamp,
-          nonce: request.nonce,
-          contentSha256: request.contentSha256,
-          signature: request.signature,
-        },
-        secret: client.secret,
-        method: request.method,
-        pathname: request.pathname,
-        contentLength: request.contentLength,
-        replayGuard,
-      });
-      return {
-        clientId: client.clientId,
-        capabilities: [
-          ...(client.capabilities.read ? (["read"] as const) : []),
-          ...(client.capabilities.write ? (["write"] as const) : []),
-        ],
-      };
-    } catch {
-      return null;
-    }
-  };
-};
-
-export type LanServerHandle = {
-  host: string;
-  port: number;
-  url: string;
-  close: (drainMs?: number) => Promise<void>;
-};
-
-const singleHeader = (
-  request: IncomingMessage,
-  name: string
-): string | null => {
-  const lowerName = name.toLowerCase();
-  let count = 0;
-  let value: string | null = null;
-  for (let index = 0; index < request.rawHeaders.length; index += 2) {
-    if (request.rawHeaders[index]!.toLowerCase() === lowerName) {
-      count += 1;
-      value = request.rawHeaders[index + 1] ?? null;
-    }
-  }
-  return count === 1 ? value : null;
-};
-
-const parseContentLength = (request: IncomingMessage): number | null => {
-  const value = singleHeader(request, "content-length");
-  if (value === null || !/^(?:0|[1-9][0-9]*)$/.test(value)) return null;
-  const parsed = Number(value);
-  return Number.isSafeInteger(parsed) ? parsed : null;
-};
-
-const sendBuffer = (
-  response: ServerResponse,
-  status: number,
-  body: Buffer,
-  contentType: string,
-  headers: Record<string, string> = {}
-): void => {
-  response.writeHead(status, {
-    "cache-control": "no-store",
-    connection: "close",
-    "content-length": String(body.length),
-    "content-type": contentType,
-    ...headers,
-  });
-  response.end(body);
-};
-
-const sendJson = (
-  response: ServerResponse,
-  status: number,
-  value: unknown
-): void =>
-  sendBuffer(
-    response,
-    status,
-    Buffer.from(`${JSON.stringify(value)}\n`, "utf8"),
-    "application/json"
-  );
-
-const sendEmpty = (
-  response: ServerResponse,
-  status: number,
-  headers: Record<string, string> = {}
-): void => {
-  response.writeHead(status, {
-    "cache-control": "no-store",
-    connection: "close",
-    "content-length": "0",
-    ...headers,
-  });
-  response.end();
-};
-
-const rejectRequest = (
-  request: IncomingMessage,
-  response: ServerResponse,
-  status: number
-): void => {
-  request.resume();
-  sendEmpty(response, status);
-};
-
-const ensureIncomingDirectory = (directory: string): void => {
-  fs.mkdirSync(directory, { recursive: true, mode: 0o700 });
-  const stats = fs.lstatSync(directory);
-  if (stats.isSymbolicLink() || !stats.isDirectory()) {
-    throw new Error("LAN incoming path must be a real directory");
-  }
-  if (process.platform !== "win32") {
-    if (
-      typeof process.getuid === "function" &&
-      stats.uid !== process.getuid()
-    ) {
-      throw new Error("LAN incoming path has a different owner");
-    }
-    fs.chmodSync(directory, 0o700);
-  }
-};
-
-const ensureTransferLocksDirectory = (directory: string): void => {
-  ensureIncomingDirectory(directory);
-};
-
-const readBody = async (
-  request: IncomingMessage,
-  expectedBytes: number,
-  maximumBytes: number
-): Promise<{ bytes: Buffer; digest: string }> => {
-  if (expectedBytes > maximumBytes) throw new Error("body-too-large");
-  const chunks: Buffer[] = [];
-  const hash = crypto.createHash("sha256");
-  let received = 0;
-  for await (const rawChunk of request) {
-    const chunk = Buffer.isBuffer(rawChunk) ? rawChunk : Buffer.from(rawChunk);
-    received += chunk.length;
-    if (received > expectedBytes || received > maximumBytes) {
-      throw new Error("body-overflow");
-    }
-    hash.update(chunk);
-    chunks.push(chunk);
-  }
-  if (received !== expectedBytes) throw new Error("body-truncated");
-  return { bytes: Buffer.concat(chunks, received), digest: hash.digest("hex") };
-};
-
-const spoolBody = async (
-  request: IncomingMessage,
-  expectedBytes: number,
-  maximumBytes: number,
-  destination: string,
-  deadlineMs: number
-): Promise<string> => {
-  if (Date.now() >= deadlineMs) throw new LanOperationTimeoutError();
-  if (expectedBytes <= 0 || expectedBytes > maximumBytes) {
-    throw new Error("body-size-invalid");
-  }
-  const descriptor = fs.openSync(
-    destination,
-    fs.constants.O_WRONLY |
-      fs.constants.O_CREAT |
-      fs.constants.O_EXCL |
-      fs.constants.O_NOFOLLOW,
-    0o600
-  );
-  const output = fs.createWriteStream("", { fd: descriptor, autoClose: true });
-  const hash = crypto.createHash("sha256");
-  let received = 0;
-  request.on("data", (rawChunk: Buffer) => {
-    const chunk = Buffer.isBuffer(rawChunk) ? rawChunk : Buffer.from(rawChunk);
-    received += chunk.length;
-    if (received > expectedBytes || received > maximumBytes) {
-      request.destroy(new Error("body-overflow"));
-      return;
-    }
-    hash.update(chunk);
-  });
-  const timeout = setTimeout(() => {
-    request.destroy(new LanOperationTimeoutError());
-  }, Math.max(1, deadlineMs - Date.now()));
-  try {
-    await pipeline(request, output);
-  } finally {
-    clearTimeout(timeout);
-  }
-  if (received !== expectedBytes) throw new Error("body-truncated");
-  return hash.digest("hex");
-};
-
-class LanOperationTimeoutError extends Error {
-  constructor() {
-    super("LAN server operation exceeded its deadline");
-  }
-}
-
-const awaitBeforeDeadline = async <T>(
-  operation: Promise<T>,
-  deadlineMs: number
-): Promise<T> => {
-  const remainingMs = deadlineMs - Date.now();
-  if (remainingMs <= 0) throw new LanOperationTimeoutError();
-  let timeout: NodeJS.Timeout | null = null;
-  try {
-    const value = await Promise.race([
-      operation,
-      new Promise<never>((_resolve, reject) => {
-        timeout = setTimeout(
-          () => reject(new LanOperationTimeoutError()),
-          remainingMs
-        );
-      }),
-    ]);
-    if (Date.now() >= deadlineMs) throw new LanOperationTimeoutError();
-    return value;
-  } finally {
-    if (timeout) clearTimeout(timeout);
-  }
-};
-
-const isStrictRequestTarget = (target: string): boolean =>
-  target.startsWith("/") &&
-  !target.includes("?") &&
-  !target.includes("#") &&
-  !target.includes("%") &&
-  !target.includes("\\") &&
-  !target.includes("//");
-
-const framingIsSafe = (request: IncomingMessage): boolean =>
-  request.httpVersionMajor === 1 &&
-  request.httpVersionMinor === 1 &&
-  request.headers["transfer-encoding"] === undefined &&
-  request.headers["content-encoding"] === undefined &&
-  request.headers.expect === undefined &&
-  parseContentLength(request) !== null;
-
-const authFields = (
-  request: IncomingMessage,
-  contentLength: number,
-  pathname: string,
-  allowPendingPairing: boolean
-): LanAuthenticationRequest | null => {
-  const clientId = singleHeader(request, "x-elc-client-id");
-  const timestamp = singleHeader(request, "x-elc-timestamp");
-  const nonce = singleHeader(request, "x-elc-nonce");
-  const contentSha256 = singleHeader(request, "x-elc-content-sha256");
-  const signature = singleHeader(request, "x-elc-signature");
-  if (
-    !clientId ||
-    !timestamp ||
-    !nonce ||
-    !contentSha256 ||
-    !signature ||
-    !SHA256.test(contentSha256)
-  ) {
-    return null;
-  }
-  return {
-    clientId,
-    timestamp,
-    nonce,
-    contentSha256,
-    signature,
-    contentLength,
-    method: request.method ?? "",
-    pathname,
-    allowPendingPairing,
-  };
-};
 
 export const startLanServer = async (
   options: LanServerOptions
